+++ src/main/java/sidkbk/celemo/controllers/OrderController.java
@@ -8,9 +8,12 @@ import org.springframework.web.bind.annotation.*;
 import sidkbk.celemo.dto.order.DeleteOrderDTO;
 import sidkbk.celemo.dto.order.OrderCreationDTO;
 import sidkbk.celemo.dto.order.OrderFoundByIdDTO;
+import sidkbk.celemo.dto.user.FindUserIdDTO;
 import sidkbk.celemo.exceptions.EntityNotFoundException;
 import sidkbk.celemo.models.Order;
 import sidkbk.celemo.services.OrderService;
+
+import java.util.List;
 
 @RestController
 @RequestMapping(value = "/api/order")
@@ -47,6 +50,17 @@ public class OrderController {
     @DeleteMapping("/delete")
     public ResponseEntity<?> deleteOrder(@Valid @RequestBody DeleteOrderDTO deleteOrderDTO) {
             return orderService.deleteOrder(deleteOrderDTO);
+    }
+
+
+    @GetMapping("/find/user-orders")
+    public ResponseEntity<?>findAllOrderForOneUser(@Valid @RequestBody FindUserIdDTO findUserIdDTO){
+        List<Order> foundOrder = orderService.findAllOrderForOneUser(findUserIdDTO);
+        if (foundOrder.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("no order found");
+        }else {
+            return ResponseEntity.ok().body(foundOrder);
+        }
     }
 }
